requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
spotipy==2.23.0
ytmusicapi==1.3.2
